Windows build fixes.
Windows doesn't have ssize_t, sadly. There's SSIZE_T, but defining an
OPENSSL_SSIZE_T seems worse than just using an int.

Reviewed-on: https://boringssl-review.googlesource.com/1352

+++ crypto/dh/dh.c
@@ -147,7 +147,7 @@ int DH_generate_key(DH *dh) {
   return DH_default_method.generate_key(dh);
 }
 
-ssize_t DH_compute_key(unsigned char *out, const BIGNUM *peers_key, DH *dh) {
+int DH_compute_key(unsigned char *out, const BIGNUM *peers_key, DH *dh) {
   if (dh->meth->compute_key) {
     return dh->meth->compute_key(dh, out, peers_key);
   }
